Crate zng_ext_hot_reload

Hot reload service.

Crate

This crate is part of the zng project.

Macros

hot_static

Declares a patchable static.

hot_static_ref

Static reference to a hot_static!.

lazy_static

Implementation of lazy_static! that supports hot reloading.

zng_hot_entry

Declare hot reload entry.

Structs

BuildArgs

Arguments for custom rebuild runners.

HOT_RELOAD

Hot reload service.

HotReloadArgs

Args for HOT_RELOAD_EVENT.

HotReloadManager

Hot reload app extension.

HotStatus

Status of a monitored dynamic library crate.

Enums

BuildError

Rebuild error.

Statics

HOT_RELOAD_EVENT

Event notifies when a new version of a hot reload dynamic library has finished rebuild and has loaded.

Attribute Macros

hot_node

Expands an impl into a UiNode impl.